Helper\UpdateHelper moved and renamed to Manager\UpdateManager
